refactor(orchestrator): extract IssueConfigStorePort
